Poll for the Heat stack state when accessing Racks
This patch will allow the Rack state to be set to 'provisioning'
and then instead of setting it to 'provisioned' right after
the Heat update_stack is called it will check the Heat stack
state everytime you access the Rack.

Once it found the Rack with state set to 'provisioning' and the Heat
update is completed, it will change the state to 'provisioned'.