Fix tests for Python 2.6
Our test cases use the `testtools` package which in turn relies on the
standard library `unittest` for the basic assert functions.

Unfortunately, Python 2.6's `unittest.TestCase` doesn't come with
`assertRegexpMatches` which we use in our tests. This caused the py26 Tox
environment to fail.

Adding `unittest2` (a backport of the 2.7 unittest for earlier versions) and
mixing it in with our base TestCase class adds the missing asserts.

Change-Id: Ifc69885421f167922b92a37f61dcab3db9665383
Signed-off-by: Tomas Sedovic <tsedovic@redhat.com>