blueprint rework-force-nodes
Implements the rework of 'force-nodes' for putting images onto specific overcloud nodes

This implements the rework of the force-nodes mechanism to use flavors and host
aggregates to specify which nodes/hosts a particular image should be launched onto
by the nova baremetal scheduler. In particular, a host aggregate is created for
each resource class and tagged with metadata (class=name). A flavor is also created
and tagged with the same metadata. All nodes in a given resource class must be placed
into the host aggregate.

This commit adds the wrapper methods around the python-novaclient to manage aggregates/flavors
in nova (baremetal). The created flavor and aggregate ID are tracked with parameters on the
resource class database model. You must specify admin credentials for undercloud nova BM
 in the nova_undercloud_config.yml file. You must specify the call_nova: "True" key to actually
call out to Nova (otherwise you just get a mock client). This last part is mainly for debugging
and will likely be removed.

Change-Id: I0ad22429957d15f70a72004333fa217667858288