blueprint rework-force-nodes - HEAT templates
Changes the yaml templates used for generation of the HEAT template to
specify the created flavor for a given resource class.

Now that we are creating aggregates/flavors for the placement of particular
images on nodes/hosts by Nova baremetal, we must specify the flavor to use
in the HEAT template.

Change-Id: Ic8130942c5c8dbea580434d678ce6b99b82a272f