Add Glance image id to `resource_classes` table
At the moment in Tuskar we have hardcoded (due to Poc) image names
in the Heat template. We should add possibility to use Glance
relevant image for a resource class during the resource class
registration (or modification) process.

This patch brings possibility to accept an image_id parameter as
part of the resource class model - column `image_id` was added to table
`resource_classes`.

Tests modified.

Closes-Bug: #1233087

Change-Id: Iad199a9adc9f276fefe4349e373b4e9da2a9fe27