Update repo for move to openstack namespace
.gitreview must be updated to move to the openstack namespace.
boto==2.5.1 being in the requirements also prevents that change from
merging.

There is no usage of the boto library anywhere in tuskar. None of the
binaries that boto ships are executed. It is unclear why boto was even
added to requirements.txt. Meanwhile we are under stricter global
requirements due to being moved into the official openstack namespace,
so having boto==2.5.1 listed here causes our gate to fail.

Change-Id: Ie54f5418dff31caf8f25c01c252839a56747518c