Revert "Support building wheels (PEP-427)"
This reverts commit 1639baab1ed3d84cd039dadd608de08ae16c6718.

The patch being reverted should not have been merged, because
setting universal to True is used to claim in-place python 2/3
compatibility in Wheel packages.

We don't support python 3 yet, so we must not claim the opposite
here (as this would harm end users as soon as we start uploading
Wheel packages to PyPi).

Please see the following links for details:
 - http://pythonwheels.com/
 - http://lists.openstack.org/pipermail/openstack-dev/2013-November/020059.html

Change-Id: If428b7e9f09374c53e8a1ea846e25287a5eca107