Use alembic for database migrations
Sqlalchemy-migrate project does not support python3 and lot of another
useful tools, which supported by alembic. On another hand, sqlalchemy-migrate
has been abandoned by its original developers, it's not actively developed
anymore, there will be no features coming, and, which is more
importantly it becomes more and more harder to keep it compatible with
the latest versions of SQLAlchemy, as well as we need to put human
resources into fixing of bugs in sqlalchemy-migrate.

Added initial migration for alembic, which based on initial
sqlalchemy-migrate migration.

Refactored code and tests to use alembic as migration api.

Co-Authored-By: Dmitry Shulyak <dshulyak@mirantis.com>

Change-Id: I214b394f35942f087cf45f5f4f4661df6459ff31