Fixes for Tuskar API Keystone authentication
Rather than duplicating the instructions as suggested in the bug, I added a link to the Keystone service configuration page instead. It offers a much more comprehensive explanation of the steps than we should.

In verifying those instructions, I had to fix code in Tuskar to get the authentication
working. The original code was copied from Ironic, however, their needs
are slightly different as they have the notion of public APIs. They use
the AdminAuthHook to facilitate admin and public access. Furthermore,
the code Tuskar did have would crash when attempting to call to_dict() on the
context, which is specifically implemented in Ironic.

Change-Id: I13c249ea02c5c74fb378faf41ada0efb55ddf162
Closes-Bug: #1284133