Remove requirements_style_check
Requirements files should not be kept in alphabetical order, because
pip does not actually have a dep solver. Order is important. This file
should never have existed. While doing this, put pbr at the top of
the file, which is where it needs to be, and remove setuptools_git
which should no longer be used.

Change-Id: I1ae61c1c5ce66ed15fbdefbee063997ea6f3172c