Temporary fix for SQLite compatibility
There is a long standing issue (see partial-bug reference to
launchpad, currently a bug with the most "heat" of the high priority
ones) that Tuskar is not compatible with recent versions of
SQLite. Currently this breaks Tuskar on distros shipping recent SQLite
(e.g. Fedora 20) and if we upgrade the CI environment, it will break
CI too.

There is already updated code in oslo-incubator that fixes this, but
updating the whole module is not entirely straightforward (it breaks
stuff and requires amending Tuskar code). Petr Blaho is already
working on this.

I know updating oslo-incubator code directly is dirty, but it's the
only option to make Tuskar work on newer SQLite until the full module
updates land (aside from compiling older SQLite or manually applying
this patch each time).

Change-Id: Ie9e0c26f2cc701fce0ec8baaa18bd22a5b6e7568
Partial-Bug: #1290849