Use Emacs-friendly file variable to set file encoding
The file encoding line serves a dual purpose: it informs both Python
and Emacs about the file encoding.

PEP 263 explains that a file with non-ASCII characters must specify an
encoding. Without an encoding line, importing the file will raise a
SyntaxError.

While Python accepts both of

  -*- encoding: <some-encoding> -*-
  -*- coding: <some-encoding> -*-

Emacs only accepts the latter since it looks for a file variable named
"coding", not "encoding". Since the Emacs specific "-*-" syntax is
already used, it seems right to use the correct variable for Emacs.

Change-Id: I72393913ba702156ab848da0835918975544b41d