Add scaling logic to the deployment plan
When adding a template to a plan, the plan will automatically wrap the
resource created for the role in a resource group and add the
appropriate count parameter for controlling how many instances to create
in that group. The composer has been updated as well to be able to
properly generate YAML for a resource definition within a resource
group.

Change-Id: I4f25b1b41b768b8ff1f319cc03b75782fd618743
Implements: blueprint tripleo-juno-tuskar-plan-rest-api