Don't allow name conflicts in environment parameters
When composing an environment with multiple parameters of the same
name, the parameter which was added last wins. However, it doesn't
make sense to have multiple parameters of the same name in the
environment object in the first place, so we should check for it and
treat it as an error.

The find_parameter_by_name function now also allows returning None
instead of raising ValueError if the parameter was not found.

Change-Id: I4e6c61527b16d8a0a7c1ca18c62327c8ee2cadbf