Remove warning about tht dir
The tht_local_dir isn't used anymore. A recent change was made to not
raise an error if the directory cannot be found. But there is still an
error message logged saying Tuskar cannot proceed, which isn't the case.
This commit removes that error message so it's not misleading to the
user.

Change-Id: Ifcd29ba8f1d7fb5061dc0ed6a613186ac89bac17