Use resource registry for seed role matching
To match a resource from the seed template to a given role, Tuskar does a
case-insensitive comparison of the role name (derived from its file name) to the
resource name.

This commit replaces that behaviour with consulting the resource-registry
environment associated with the seed template.

Matching roles to the associated seed resources is now done through the custom
types defined in the resource-registry. The resource names no longer hold any
special meaning.

This is a breaking change as it adds a new required parameter to
`tuskar-load-roles` when `--master-seed` is used.

Change-Id: I6d2f44332404fd9cab846df1256a2804dce42d70