WIP: Support non-role resources in resource registry
This change allows for resources utilized in the master seed that are
specified in the resource registry, but are not a provider resource to
be stored and saved to disk. There is no additional information required
to pass as all the data is stored while loading roles.

--- Will snip later:
This is a proof of concept for review, but probably needs to be modified
with a schema change to remove some limitations, which are:
- Entries in the resource registry are assumed to not be using the
  Tuskar namespace.
- Currently multiple plans with the same filename are going to conflict.

Change-Id: Ia15c0fdad0c036564b81f664e49fca280bc7ce7e