WIP: Support non-role resources in resource registry
This change allows for resources utilized in the master seed that are
specified in the resource registry, but are not a provider resource to
be stored and saved to disk. There is no additional information required
to pass as all the data is stored while loading roles.

Previously tuskar always assumed that there was a 1:1 correspondence
between the final resource registry and the number of roles, but this is
no longer the case.

Change-Id: Ia15c0fdad0c036564b81f664e49fca280bc7ce7e