Adds tuskar-load-role to load a single role and associated extra-data
This adds a stand-alone tuskar-load-role command which takes a name
and path to the main file defining this role. If no name is provided
this is deduced from the path (filename) as currently occurs with
tuskar-load-roles. If a path is not specified then a SystemExit
occurs.

Optionally you may specify the extra-data files that the given role
uses as multiple '--extra-data' arguments to tuskar-load-role.

Change-Id: I9c143afb52c43e4258c3a6797a9707c08d8dcdaf