Make attribute name match resource name
In aa754741a32426f6ef8f867cf4968e30696ed380, the -servers suffix was
dropped, so the name used from the resource group (for example in the
seed template) needs to also drop the suffix.

Have grepped the source for any remaining "-server" strings, so I think
this completes the transition.

Change-Id: I0366953e1186c0aeddf1cfbf5814c6fdf4312279